+++ CHANGELOG.md
@@ -11,6 +11,10 @@ WIP
 * {GollyUtils::Callbacks Callbacks} can now be defined in modules.
 * Added {GollyUtils::Callbacks::ClassMethods#callbacks} and {GollyUtils::Callbacks::ModuleMethods#callbacks} to provide a list of all defined and inherited callbacks.
 * {GollyUtils::Callbacks::InstanceMethods#run_callbacks Callbacks#run_callbacks} now accepts options which currently allows for callback arguments to be specified.
+* {GollyUtils::Callbacks::InstanceMethods#run_callback Callbacks#run_callback} and {GollyUtils::Callbacks::InstanceMethods#run_callbacks Callbacks#run_callbacks}
+  now accept an option to pass arguments to callback functions.
+* {GollyUtils::Callbacks::InstanceMethods#run_callback Callbacks#run_callback} and {GollyUtils::Callbacks::InstanceMethods#run_callbacks Callbacks#run_callbacks}
+  now accept an option to provide a context for callback function execution.
 * Added Ruby extensions:
   * {Enumerable#frequency_map}
 * Added test helpers:
@@ -32,7 +36,7 @@ WIP
 #### Modified
 * Finished the experiment that was {GollyUtils::AttrDeclarative}. No mucking around with blocks, eager/lazy evaluation anymore; now it's all values.
 * {GollyUtils::Delegator#respond_to? Delegator#respond_to?} now also indicates `true` for local methods (in addition to delegatable).
-* {GollyUtils::Callbacks::InstanceMethods#run_callback Callbacks#run_callback} now only runs a single callback, and accepts arguments to pass to each callback.
+* {GollyUtils::Callbacks::InstanceMethods#run_callback Callbacks#run_callback} now only runs a single callback.
 * Renamed dir: `golly-utils/test` => `golly-utils/testing`.
 * Renamed dir: `golly-utils/test/spec` => `golly-utils/testing/rpsec`.
 * Renamed module: `GollyUtils::TestHelpers` => `GollyUtils::Testing::Helpers`